Fixed two more warnings.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@2462 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/smokeparticles.cpp
@@ -206,7 +206,7 @@ bool idSmokeParticles::EmitSmoke( const idDeclParticle *smoke, const int systemS
 			if ( nowCount >= stage->totalParticles ) {
 				nowCount = stage->totalParticles-1;
 			}
-			prevCount = floor( ((float)( deltaMsec - USERCMD_MSEC ) / finalParticleTime) * stage->totalParticles );
+			prevCount = static_cast<int>(floor( ((float)( deltaMsec - USERCMD_MSEC ) / finalParticleTime) * stage->totalParticles ));
 			if ( prevCount < -1 ) {
 				prevCount = -1;
 			}
